backend/requirements.txt
# Core Requirements
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
geoalchemy2==0.14.7

# AI/ML
google-generativeai==0.3.2
tensorflow==2.15.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3

# Geospatial
shapely==2.0.2
geopy==2.4.1
pyproj==3.6.1

# Utilities
python-multipart==0.0.6
httpx==0.26.0
python-jose==3.3.0
passlib==1.7.4

# Development
pytest==7.4.4
pytest-asyncio==0.23.5
requests==2.31.0